app/requirements.txt
django==5.0.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
pytest-django==4.8.0
pytest==8.0.0
celery[redis]==5.3.6
django-redis==5.4.0
django-cachalot==2.6.2